perf: split the KeepElements barrel so the shell stops pulling all 34 elements
#813 step 2, and the one that mattered. Eager closure 1,465.3 -> 1,222.5 kB raw,
371.4 -> 332.9 kB gzip: **-242.8 kB raw, -16.6 %**. Cumulative with step 1, 1,769.9 ->
1,222.5, **-30.9 %**.

Every wrapper used to be declared in one module, so importing `KeepTooltip` pulled in
Monaco's wrapper, the data table, the source editor and every WebAwesome component any of
the 34 reach. Twelve modules on the critical path did that. The 777.3 kB `useSlot` chunk
step 1 left behind was the result, and it is gone.

The declarations move to one file each under `keep-elements/react/`, and `KeepElements.tsx`
becomes a pure re-export barrel. That is deliberate: the ~50 route-local importers and
every test keep their import path, and inside a route chunk the whole set is wanted anyway.
Only the eager modules import `./react/<Name>` directly.

**Twelve, not the eight the issue listed.** Rather than trust the list I walked static
imports from `src/index.tsx`, following `from '...'` and never `import(`. That found
`Views.tsx` (which step 1 itself made an importer), `App.tsx`, and three reached through
`QuickConfigFormContainer` -- and it cleared `NetworkErrorDialog.tsx`, which the issue
lists but which nothing statically reaches. Eager source modules fall 172 -> 125.

The wyw exclude widens from `**/components/keep-elements/*.ts` to
`**/components/keep-elements/**` in both bundler configs. A single `*` does not cross a
directory boundary, so `react/` would have fallen back into wyw's scope, and those modules
import element classes using `accessor` -- the exact combination #747 found oxc
mis-desugars, which fails at module load rather than at build. Nothing under
`keep-elements/` contains Linaria, `KeepElements.tsx` included, so widening costs nothing.

`test/decorator-config.test.ts` asserted the old glob literally and went red, which is the
guard behaving correctly rather than an obstacle. Updated, with the subdirectory reason
recorded so the next person does not narrow it back.

Coverage on `keep-elements/**` *improves*, 86.7 -> 88.2 lines: all 34 new wrappers are
exercised through the barrel and land at 100 %. Floors 83/83/82/67, unchanged.

Budget tightened 1,494.7 -> 1,246.9 kB raw.

**What is still eager.** `KeepMobileHeader-*.js` at 543.6 kB is the WebAwesome subset the
remaining eager elements reach -- tooltip, footer, alert, button, dropdown, drawer, tree,
checkbox and the mobile header. Step 3 makes `QuickConfigFormContainer` conditional, which
drops six of those from the critical path.

typecheck 0, lint 0, build clean, budget green, 111 files / 1355 tests.
