+++ bundle-budget.json
@@ -1,9 +1,9 @@
 {
-  "raw": 1494655,
-  "gzip": 378800,
+  "raw": 1246944,
+  "gzip": 339546,
   "measured": {
-    "raw": 1465348,
-    "gzip": 371372
+    "raw": 1222494,
+    "gzip": 332888
   },
   "headroom": 0.02
 }
