+++ src/AppShell.tsx
@@ -21,7 +21,9 @@ import ProfileMenuDialog from './components/sidenav/ProfileMenuDialog';
 import Notification from './components/alerts/Notification';
 import SideNav from './components/sidenav/SideNav';
 import ProfileMenu from './components/sidenav/ProfileMenu';
-import { KeepFooter, KeepMobileHeader, KeepTooltip } from './components/keep-elements/KeepElements';
+import { KeepFooter } from './components/keep-elements/react/KeepFooter';
+import { KeepMobileHeader } from './components/keep-elements/react/KeepMobileHeader';
+import { KeepTooltip } from './components/keep-elements/react/KeepTooltip';
 import { applyTheme } from './services/theme-service';
 import { AppState } from './store';
 import { getTheme, switchTheme } from './store/styles/action';
