+++ src/Views.tsx
@@ -11,7 +11,8 @@ import { useSelector } from 'react-redux';
 import { AppState } from './store';
 import { setLoading } from './store/loading/action';
 import BreadcrumbRouter from './components/routers/BreadcrumbRouter';
-import { KeepPageLoading, KeepPageRouters } from './components/keep-elements/KeepElements';
+import { KeepPageLoading } from './components/keep-elements/react/KeepPageLoading';
+import { KeepPageRouters } from './components/keep-elements/react/KeepPageRouters';
 import { fetchScopes, fetchKeepPermissions } from './store/databases/action';
 import { NavigationGuardProvider } from './components/navigation/NavigationGuardContext';
 import QuickConfigFormContainer from './components/database/QuickConfigFormContainer';
