+++ src/components/keep-elements/react/KeepAppStatus.ts
@@ -1,0 +1,15 @@
+/* ========================================================================== *
+ * Copyright (C) 2026 HCL America Inc.                                        *
+ * All rights reserved.                                                       *
+ * Licensed under Apache 2 License.                                           *
+ * ========================================================================== */
+
+import React from 'react';
+import { createComponent } from '@lit/react';
+import AppStatus from '../keep-app-status';
+
+export const KeepAppStatus = createComponent({
+  tagName: 'keep-app-status',
+  elementClass: AppStatus,
+  react: React
+});
