+++ src/components/keep-elements/react/KeepZeroResults.ts
@@ -1,0 +1,15 @@
+/* ========================================================================== *
+ * Copyright (C) 2026 HCL America Inc.                                        *
+ * All rights reserved.                                                       *
+ * Licensed under Apache 2 License.                                           *
+ * ========================================================================== */
+
+import React from 'react';
+import { createComponent } from '@lit/react';
+import ZeroResults from '../keep-zero-results';
+
+export const KeepZeroResults = createComponent({
+  tagName: 'keep-zero-results',
+  elementClass: ZeroResults,
+  react: React
+});
